requirements.txt
streamlit
google-adk>=1.18.0
nest_asyncio